# perf: combine multiple git add calls into single operation