--- a/lib/gitWorkflow.js
+++ b/lib/gitWorkflow.js
@@ -275,9 +275,12 @@ export class GitWorkflow {
   async applyModifications(ctx) {
     debugLog('Adding task modifications to index...')
 
+    // Collect all addable files from all chunks first
+    const allAddableFiles = []
+
     // `matchedFileChunks` includes staged files that lint-staged originally detected and matched against a task.
     // Add only these files so any 3rd-party edits to other files won't be included in the commit.
-    // These additions per chunk are run "serially" to prevent race conditions.
+    // Files are processed serially to avoid race conditions, then added in a single git command.
     // Git add creates a lockfile in the repo causing concurrent operations to fail.
     for (const files of this.matchedFileChunks) {
       const accessCheckedFiles = await Promise.allSettled(
@@ -295,7 +298,12 @@ export class GitWorkflow {
         r.status === 'fulfilled' ? [r.value] : []
       )
 
-      await this.execGit(['add', '--', ...addableFiles])
+      allAddableFiles.push(...addableFiles)
+    }
+
+    // Add all files in a single git add command instead of multiple calls
+    if (allAddableFiles.length > 0) {
+      await this.execGit(['add', '--', ...allAddableFiles])
     }
 
     debugLog('Done adding task modifications to index!')