@iiroj 2025-06-22
Hello,

the purpose of this batching is to avoid hitting the maximum command string length of terminals; it should be mostly a problem on Windows, but on other OSes there shouldn't typically be more than a single batch.

How did you end up spending time on this, did you observe a real-world issue where there are multiple batches while there shouldn't have to be?