@outslept 2025-06-22
Good point! Combining chunks could exceed Windows limits. Windows strikes again! 😅