@iiroj 2025-06-22
Are you by chance using AI to search for performance improvements in this project? If so, please don't, it just wastes everyone's time.

I do appreciate simple and valid fixes, but this PR gives me a bad vibe because while seemingly valid and with tests, it misses the point of the existing functionality and would effectively just undo a previous fix that is there for a reason.